fix(solver): split oversized files under the 2000-line cap (§19)
core.rs (2011 → 1949 lines): move ContentAddressedDefIds into the
existing core/ submodule directory as core/content_addressed.rs and
re-export it from core.rs.

def_tests.rs (2145 → 1875 lines): move the register_type_to_def
priority tests into def_tests_parts/type_to_def.rs (included via
include!). Adds one new test —
register_type_to_def_uncontested_interface_stays_as_interface — which
proves the TypeAlias-wins rule only fires when alias and non-alias defs
actually compete for the same TypeId; an interface with its own unique
TypeId is preserved without displacement.
